Move validators from auth reducer
