+++ src/renderer/actions/index.ts
@@ -2,3 +2,4 @@ export * from "./register";
 export * from "./deposit";
 export * from "./auth";
 export * from "./notification";
+export * from "./validator";
