+++ src/renderer/reducers/validators.ts
@@ -1,6 +1,7 @@
 import {Action} from "redux";
+import {ILoadValidators} from "../actions";
 
-import {ILoadedValidatorBalanceAction} from "../actions/validator";
+import {ILoadedValidatorBalanceAction} from "../actions";
 import {ValidatorActionTypes} from "../constants/action-types";
 import {IValidator} from "../containers/Dashboard/DashboardContainer";
 
@@ -17,6 +18,14 @@ export const validatorsReducer = (
 ): IValidatorState => {
     let payload: any;
     switch (action.type) {
+        case ValidatorActionTypes.LOAD_VALIDATORS:
+            // eslint-disable-next-line no-case-declarations
+            const validatorMap: IValidatorState = {};
+            (action as ILoadValidators).payload.forEach((v: IValidator) => {
+                validatorMap[v.publicKey] = v;
+            });
+
+            return validatorMap;
         case ValidatorActionTypes.LOADED_VALIDATOR_BALANCE:
             payload = (action as ILoadedValidatorBalanceAction).payload;
             return {
